+++ packages/react/src/index.ts
@@ -1,3 +1,6 @@
+export * from "./types";
 export * from "./hooks/index-external";
+export * from "./parts";
 export * from "./presets";
 export * from "./components";
+