+++ packages/react/src/parts/Answer.tsx
@@ -1,0 +1,16 @@
+
+export type AnswerProps = Readonly<{
+  text: string
+}>
+
+export function Answer(props: AnswerProps) {
+  const {
+    text,
+  } = props;
+
+  return (
+    <div className="feelback-a">
+      <span className="feelback-text">{text}</span>
+    </div>
+  );
+}