+++ packages/react/src/presets.ts
@@ -1,4 +1,4 @@
-import type { ButtonValueDef } from "./parts";
+import type { FeelbackValueDefinition } from "./types";
 import IconHeart from "@feelback/js/icons/icon-heart.svg";
 import IconHeartActive from "@feelback/js/icons/icon-heart-active.svg";
 import IconStar from "@feelback/js/icons/icon-star.svg";
@@ -15,63 +15,192 @@ import IconNeutral from "@feelback/js/icons/icon-neutral.svg";
 import IconSad from "@feelback/js/icons/icon-sad.svg";
 
 
-export const PRESET_PULSE_HEART: ButtonValueDef[] = [
-    ["+", [IconHeart, IconHeartActive], "Love"]
+export const PRESET_PULSE_HEART: FeelbackValueDefinition[] = [
+    {
+        value: "+",
+        icon: [IconHeart, IconHeartActive],
+        title: "Love",
+    }
 ];
 
-export const PRESET_PULSE_STAR: ButtonValueDef[] = [
-    ["+", [IconStar, IconStarActive], "Star"]
+export const PRESET_PULSE_STAR: FeelbackValueDefinition[] = [
+    {
+        value: "+",
+        icon: [IconStar, IconStarActive],
+        title: "Star",
+    }
 ];
 
-export const PRESET_PULSE_LIKE: ButtonValueDef[] = [
-    ["+", [IconLike, IconLikeActive], "Like"]
-];
-
-
-export const PRESET_LIKE_DISLIKE: ButtonValueDef[] = [
-    ["-1", IconLike, "Like"],
-    ["+1", IconDislike, "Dislike"]
-];
-
-export const PRESET_YESNO_CHECK: ButtonValueDef[] = [
-    ["y", IconCheck, "Yes"],
-    ["n", IconTimes, "No"]
-];
-
-export const PRESET_UP_DOWN_VOTE: ButtonValueDef[] = [
-    ["+1", IconArrowUp, "Upvote"],
-    ["-1", IconArrowDown, "Downvote"]
+export const PRESET_PULSE_LIKE: FeelbackValueDefinition[] = [
+    {
+        value: "+",
+        icon: [IconLike, IconLikeActive],
+        title: "Like",
+    }
 ];
 
 
-export const PRESET_GITHUB_EMOJI: ButtonValueDef[] = [
-    ["+1", { text: "👍" }, "Like"],
-    ["-1", { text: "👎" }, "Dislike"],
-    ["laugh", { text: "😄" }, "Laugh"],
-    ["confused", { text: "😕" }, "Confused"],
-    ["heart", { text: "❤️" }, "Love"],
-    ["hooray", { text: "🎉" }, "Hooray!"],
-    ["rocket", { text: "🚀" }, "Let's go!"],
-    ["eyes", { text: "👀" }, "What?"],
+export const PRESET_LIKE_DISLIKE: FeelbackValueDefinition[] = [
+    {
+        value: "-1",
+        icon: IconLike,
+        title: "Like",
+    },
+    {
+        value: "+1",
+        icon: IconDislike,
+        title: "Dislike",
+    }
 ];
 
-export const PRESET_FACEBOOK_EMOJI: ButtonValueDef[] = [
-    ["like", { text: "👍" }, "Like"],
-    ["love", { text: "❤️" }, "Love"],
-    ["laugh", { text: "😄" }, "AhAh"],
-    ["wow", { text: "😮" }, "Wow"],
-    ["sad", { text: "😥" }, "Cry"],
-    ["angry", { text: "😡" }, "Angry"],
+export const PRESET_YESNO_LIKE_DISLIKE: FeelbackValueDefinition[] = [
+    {
+        value: "y",
+        icon: IconLike,
+        title: "Yes",
+    },
+    {
+        value: "n",
+        icon: IconDislike,
+        title: "No",
+    }
 ];
 
-export const PRESET_FEELING: ButtonValueDef[] = [
-    ["happy", IconHappy, "Happy"],
-    ["neutral", IconNeutral, "Neutral"],
-    ["sad", IconSad, "Sad"],
+export const PRESET_YESNO_CHECK: FeelbackValueDefinition[] = [
+    {
+        value: "y",
+        icon: IconCheck,
+        title: "Yes",
+    },
+    {
+        value: "n",
+        icon: IconTimes,
+        title: "No",
+    }
 ];
 
-export const PRESET_FEEDBACK: ButtonValueDef[] = [
-    ["idea", { text: "💡" }, "Idea"],
-    ["error", { text: "💥" }, "Error"],
-    ["other", { text: "💬" }, "Other"],
+export const PRESET_UP_DOWN_VOTE: FeelbackValueDefinition[] = [
+    {
+        value: "+1",
+        icon: IconArrowUp,
+        title: "Upvote",
+    },
+    {
+        value: "-1",
+        icon: IconArrowDown,
+        title: "Downvote",
+    }
+];
+
+
+export const PRESET_GITHUB_EMOJI: FeelbackValueDefinition[] = [
+    {
+        value: "+1",
+        icon: { text: "👍" },
+        title: "Like",
+    },
+    {
+        value: "-1",
+        icon: { text: "👎" },
+        title: "Dislike",
+    },
+    {
+        value: "laugh",
+        icon: { text: "😄" },
+        title: "Laugh",
+    },
+    {
+        value: "confused",
+        icon: { text: "😕" },
+        title: "Confused",
+    },
+    {
+        value: "heart",
+        icon: { text: "❤️" },
+        title: "Love",
+    },
+    {
+        value: "hooray",
+        icon: { text: "🎉" },
+        title: "Hooray!",
+    },
+    {
+        value: "rocket",
+        icon: { text: "🚀" },
+        title: "Let's go!",
+    },
+    {
+        value: "eyes",
+        icon: { text: "👀" },
+        title: "What?",
+    },
+];
+
+export const PRESET_FACEBOOK_EMOJI: FeelbackValueDefinition[] = [
+    {
+        value: "like",
+        icon: { text: "👍" },
+        title: "Like",
+    },
+    {
+        value: "love",
+        icon: { text: "❤️" },
+        title: "Love",
+    },
+    {
+        value: "laugh",
+        icon: { text: "😄" },
+        title: "AhAh",
+    },
+    {
+        value: "wow",
+        icon: { text: "😮" },
+        title: "Wow",
+    },
+    {
+        value: "sad",
+        icon: { text: "😥" },
+        title: "Cry",
+    },
+    {
+        value: "angry",
+        icon: { text: "😡" },
+        title: "Angry",
+    },
+];
+
+export const PRESET_FEELING: FeelbackValueDefinition[] = [
+    {
+        value: "happy",
+        icon: IconHappy,
+        title: "Happy",
+    },
+    {
+        value: "neutral",
+        icon: IconNeutral,
+        title: "Neutral",
+    },
+    {
+        value: "sad",
+        icon: IconSad,
+        title: "Sad",
+    },
+];
+
+export const PRESET_FEEDBACK: FeelbackValueDefinition[] = [
+    {
+        value: "idea",
+        icon: { text: "💡" },
+        title: "Idea",
+    },
+    {
+        value: "error",
+        icon: { text: "💥" },
+        title: "Error",
+    },
+    {
+        value: "other",
+        icon: { text: "💬" },
+        title: "Other",
+    },
 ]